services/backend/requirements.txt
aerich==0.7.1
asyncpg==0.29.0
passlib==1.7.4
fastapi==0.88.0
tortoise-orm==0.19.2
uvicorn==0.20.0